JAMES-1818 Move TextExtractor from store to API
