+++ mailbox/store/src/test/java/org/apache/james/mailbox/store/extractor/DefaultTextExtractorTest.java
@@ -19,11 +19,13 @@
 
 package org.apache.james.mailbox.store.extractor;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import java.io.InputStream;
 
+import org.apache.james.mailbox.extractor.TextExtractor;
 import org.junit.Before;
 import org.junit.Test;
-import static org.assertj.core.api.Assertions.assertThat;
 
 public class DefaultTextExtractorTest {
     private TextExtractor textExtractor;
