+++ mailbox/tika/src/test/java/org/apache/james/mailbox/tika/extractor/TikaTextExtractorTest.java
@@ -19,11 +19,11 @@
 
 package org.apache.james.mailbox.tika.extractor;
 
-import java.io.InputStream;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.apache.james.mailbox.store.extractor.TextExtractor;
+import java.io.InputStream;
+
+import org.apache.james.mailbox.extractor.TextExtractor;
 import org.junit.Before;
 import org.junit.Test;
